+++ src/config/routes/component-routes.js
@@ -1,6 +1,7 @@
+//External
+const componentRouter = require("express").Router();
+//controllers
 const {
-  //router
-  componentRouter,
   //checks
   checkBodyFieldsAddComponent,
   checkBodyFieldsUpdateComponent,
