+++ src/services/component/get-all-like-stock-min-max.js
@@ -1,0 +1,129 @@
+//Externals
+const { Op } = require('sequelize');
+//Models
+const { Component } = require('../../models/sequelize/component');
+//Enums
+const { statusName } = require('../../enums/database/status');
+const { checkErrors } = require('../../helpers/sequelize/errors');
+const {
+  checkOrderBy,
+  checkOrderAt,
+} = require('../../helpers/pagination/components/component');
+const { paginationNameValueError } = require('../../enums/pagination/errors');
+//const
+const ORDER_BY_NAME_VALUE_ERROR =
+  paginationNameValueError.ORDER_BY_NAME_VALUE_ERROR;
+const ORDER_AT_NAME_VALUE_ERROR =
+  paginationNameValueError.ORDER_AT_NAME_VALUE_ERROR;
+//componentss
+let componentList;
+//params
+let queryStrParams;
+let stockMin;
+let stockMax;
+//pagination
+let pageSizeNro = 30;
+let pageNro = 0;
+let orderBy;
+let orderAt;
+let order;
+let msg;
+
+
+
+/**
+ * @description get all paginated components list according to its stock min and max from the database
+ * @param {any} req any type
+ * @param {any} res any type
+ * @returns a json object with the transaction performed
+ * @example
+ */
+const getAllComponentLikeStockMinMaxService = async (req, res) => {
+  try {
+    componentList = null;
+    msg = null;
+    queryStrParams = null;
+    stockMin = 0;
+    stockMax = 100;
+    //Pagination
+    pageSizeNro = 5;
+    pageNro = 0;
+    orderBy = 'id';
+    orderAt = 'ASC';
+    msg = null;
+
+    //-- start with querys params and pagination  ---
+    queryStrParams = req.query;
+
+    if (queryStrParams != (null && undefined)) {
+      //stock
+      stockMin = queryStrParams.stockMin
+        ? parseInt(queryStrParams.stockMin)
+        : stockMin;
+      stockMax = queryStrParams.stockMax
+        ? parseInt(queryStrParams.stockMax)
+        : stockMax;
+      //Pagination
+      pageSizeNro = queryStrParams.limit
+        ? parseInt(queryStrParams.limit)
+        : pageSizeNro;
+      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
+      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
+      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
+    }
+
+    orderBy = await checkOrderBy(orderBy);
+
+    if (orderBy == (null || undefined)) {
+      return ORDER_BY_NAME_VALUE_ERROR;
+    }
+
+    orderAt = await checkOrderAt(orderAt);
+
+    if (orderAt == (null || undefined)) {
+      return ORDER_AT_NAME_VALUE_ERROR;
+    }
+
+    order = [[orderBy, orderAt]];
+
+    //-- end with pagination  ---
+
+    if (Component != (null && undefined)) {
+      await Component.findAll({
+        attributes: {},
+        where: {
+          stock: {
+            [Op.gte]: stockMin,
+            [Op.lte]: stockMax,
+          },
+        },
+        limit: pageSizeNro,
+        offset: pageNro,
+        order: order,
+        raw: true,
+        nest: true,
+      })
+        .then(async (componentItems) => {
+          componentList = componentItems;
+        })
+        .catch(async (error) => {
+          msg = `Error in getAllComponentLikeStockMinMaxService() function when trying to get all paginated component by stock min and stock max. Caused by ${error}`;
+          console.log(msg);
+
+          componentList = await checkErrors(error, error.name);
+        });
+    } else {
+      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
+    }
+  } catch (error) {
+    msg = `Error in getAllComponentLikeStockMinMaxService() function. Caused by ${error}`;
+    console.log(msg);
+
+    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
+  }
+  return componentList;
+};
+
+module.exports = {
+    getAllComponentLikeStockMinMaxService
+}
