+++ src/services/component/get-all-like-stock.js
@@ -1,0 +1,125 @@
+//Externals
+const { Op } = require('sequelize');
+//Models
+const { Component } = require('../../models/sequelize/component');
+//Enums
+const { statusName } = require('../../enums/database/status');
+const { checkErrors } = require('../../helpers/sequelize/errors');
+const {
+  checkOrderBy,
+  checkOrderAt,
+} = require('../../helpers/pagination/components/component');
+const { paginationNameValueError } = require('../../enums/pagination/errors');
+//const
+const ORDER_BY_NAME_VALUE_ERROR =
+  paginationNameValueError.ORDER_BY_NAME_VALUE_ERROR;
+const ORDER_AT_NAME_VALUE_ERROR =
+  paginationNameValueError.ORDER_AT_NAME_VALUE_ERROR;
+//componentss
+let componentList;
+//params
+let queryStrParams;
+let params;
+let stockParam;
+//pagination
+let pageSizeNro = 30;
+let pageNro = 0;
+let orderBy;
+let orderAt;
+let order;
+let msg;
+
+/**
+ * @description get all paginated components list according to its stock from the database
+ * @param {any} req any type
+ * @param {any} res any type
+ * @returns a json object with the transaction performed
+ * @example
+ */
+const getAllComponentLikeStockService = async (req, res) => {
+  try {
+    componentList = null;
+    queryStrParams = null;
+    stockParam = null;
+    //Pagination
+    pageSizeNro = 5;
+    pageNro = 0;
+    orderBy = 'id';
+    orderAt = 'ASC';
+    msg = null;
+
+    //-- start with params ---
+    params = req.params;
+
+    if (params != (null && undefined)) {
+      stockParam = params.stock ? parseInt(params.stock) : stockParam;
+    }
+    //-- end with params ---
+
+    //-- start with querys params and pagination  ---
+    queryStrParams = req.query;
+
+    if (queryStrParams != (null && undefined)) {
+      //Pagination
+      pageSizeNro = queryStrParams.limit
+        ? parseInt(queryStrParams.limit)
+        : pageSizeNro;
+      pageNro = queryStrParams.page ? parseInt(queryStrParams.page) : pageNro;
+      orderBy = queryStrParams.orderBy ? queryStrParams.orderBy : orderBy;
+      orderAt = queryStrParams.orderAt ? queryStrParams.orderAt : orderAt;
+    }
+
+    orderBy = await checkOrderBy(orderBy);
+
+    if (orderBy == (null || undefined)) {
+      return ORDER_BY_NAME_VALUE_ERROR;
+    }
+
+    orderAt = await checkOrderAt(orderAt);
+
+    if (orderAt == (null || undefined)) {
+      return ORDER_AT_NAME_VALUE_ERROR;
+    }
+
+    order = [[orderBy, orderAt]];
+
+    //-- end with pagination  ---
+
+    if (Component != (null && undefined)) {
+      await Component.findAll({
+        attributes: {},
+        where: {
+          stock: {
+            [Op.eq]: stockParam,
+          },
+        },
+        limit: pageSizeNro,
+        offset: pageNro,
+        order: order,
+        raw: true,
+        nest: true,
+      })
+        .then(async (componentItems) => {
+          componentList = componentItems;
+        })
+        .catch(async (error) => {
+          msg = `Error in getAllComponentLikeStockService() function when trying to get all paginated component by stock. Caused by ${error}`;
+          console.log(msg);
+
+          componentList = await checkErrors(error, error.name);
+        });
+    } else {
+      componentList = await checkErrors(null, statusName.CONNECTION_REFUSED);
+    }
+  } catch (error) {
+    msg = `Error in getAllComponentLikeStockService() function. Caused by ${error}`;
+    console.log(msg);
+
+    componentList = await checkErrors(error, statusName.CONNECTION_ERROR);
+  }
+  return componentList;
+};
+
+module.exports = {
+    getAllComponentLikeStockService
+}
